refactor: route cost dashboard through app-level hash handling
